+++ android/src/main/kotlin/com/flutter_rust_bridge/xue_hua_media_compression/XueHuaMediaCompressionPlugin.kt
@@ -1,0 +1,21 @@
+package com.flutter_rust_bridge.xue_hua_media_compression
+
+import android.content.Context
+import io.flutter.embedding.engine.plugins.FlutterPlugin
+
+class XueHuaMediaCompressionPlugin : FlutterPlugin {
+    companion object {
+        init {
+            System.loadLibrary("xue_hua_media_compression")
+        }
+
+        @JvmStatic
+        external fun initAndroid(context: Context)
+    }
+
+    override fun onAttachedToEngine(binding: FlutterPlugin.FlutterPluginBinding) {
+        initAndroid(binding.applicationContext)
+    }
+
+    override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) {}
+}